Refactor kani-verifier
